requirements.txt
boto3~=1.35.25
botocore~=1.35.25
requests~=2.32.3
python-dotenv~=1.0.1
pyairtable~=2.3.3
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0